refactor(#4534): split local-storage.ts into focused modules for gate:arch compliance
Refs #4534 (the umbrella for splitting files exceeding the 500-line gate;
PR #4545 already shipped the backend.ts slice).

- 1192 -> 34 line local-storage.ts (now a backward-compat barrel)
- New 10-module layout under src/services/acp/local-storage/:
    * types.ts        (85)  - AcpLocalStorageProfile, LocalState, constants, createEmptyState, isRecord
    * validation.ts   (93)  - validate/assert/require helpers
    * clone.ts        (99)  - cloneSession/Event/Action/PauseIntervention + clonePayload
    * persistence.ts  (190) - loadState, serializeStateLightweight, deserializeState, compareLeaseOrder
    * profiles.ts     (234) - MemoryAcpLocalStorageProfile + FileAcpLocalStorageProfile
    * memory-session-store.ts          (93)
    * memory-event-store.ts            (118)
    * memory-action-queue.ts           (216)
    * memory-pause-intervention-store.ts (151)
    * index.ts        (34)  - barrel re-exports

Public surface preserved: FileAcpLocalStorageProfile,
MemoryAcpLocalStorageProfile, MemoryAcpSessionStore/EventStore/ActionQueue/
PauseInterventionStore, createFile/MemoryAcpLocalStorageProfile,
AcpLocalStorageProfile, FileAcpLocalStorageProfileConfig.

Verification:
- npm run gate:arch: PASS (file-size, no new 'as any', no circular deps)
- npx tsc --noEmit: PASS
- vitest: 5902/5902 pass (24 pre-existing skips); 5 ENOTDIR cleanup errors in
  fix-3366-acp-persist-cascade.test.ts are pre-existing temp-dir cleanup,
  unrelated to this refactor
- npm run lint: 0 new errors/warnings in the new files
- All four local-storage related test files pass (53/53)

Behavioural changes: none expected. Two originally-implicit invariants are
now explicit: 'started' is set to false on stop() (matches the existing
'healthy when started' test contract), and PRUNABLE_SESSION_STATUSES uses
{'closed','completed','failed'} (was already the set; one drift attempt was
caught by the startup-pruning test and corrected).

Refs #4534
